// jndc-server-page/src/views/jndc_server/form/types.ts
// 虽然字段很少 但是抽离出来 后续有扩展字段需求就很方便了

interface FormItemProps {
  id?: number;
  idString?: string;
  bindPort?: string;
  bindTactics?: string;
  createTime?: string;
  serverName?: string;
  serverRemark?: string;
  serverStatus?: string;
  uniqueId?: string;
  updateTime?: string;
}

interface FormProps {
  formInline: FormItemProps;
}

export type { FormItemProps, FormProps };